Move ActionTypes and ActionManager into action-manager module
