+++ settings.gradle.kts
@@ -2,6 +2,7 @@ rootProject.name = "MonoSketch"
 
 val moduleMap = mapOf(
     "app" to "app",
+    "action-manager" to "libs/action-manager",
     "build-environment" to "libs/build-environment",
     "commons" to "libs/commons",
     "export-shapes-modal" to "libs/export-shapes-modal",
